Migrate test case
